fix types 2

+++ Api/Models/MovieRecognitionStatisticsDto.cs
@@ -5,11 +5,11 @@ namespace Api.Models;
 public class MovieRecognitionStatisticsDto
 {
 	[JsonPropertyName("total_recognized")]
-	public required int TotalRecognized { get; set; }
+	public required long TotalRecognized { get; set; }
 	
 	[JsonPropertyName("correctly_recognized")]
-	public required int CorrectlyRecognized { get; set; }
+	public required long CorrectlyRecognized { get; set; }
 	
 	[JsonPropertyName("incorrectly_recognized")]
-	public required int IncorrectlyRecognized { get; set; }
+	public required long IncorrectlyRecognized { get; set; }
 }
